# schemas/openapi/llama-server.openapi.yaml
openapi: 3.1.0
info:
  title: llama.cpp-stack API Surface
  version: 0.1.0
  description: |
    Public gateway contract for llama.cpp-stack. The gateway exposes OpenAI-compatible
    /v1/* endpoints plus /health, while hiding llama.cpp router management/native
    endpoints such as /slots, /metrics, /props, /models/load, /models/unload, and
    /completion.
servers:
  - url: https://llamacpp-stack.vex9z7.com
    description: Current deployed stack
  - url: http://127.0.0.1:8090
    description: Local gateway development
paths:
  /health:
    get:
      summary: Gateway health check
      operationId: getHealth
      responses:
        '200':
          description: Server is ready
          content:
            application/json:
              schema:
                $ref: '../json/health-response.schema.json'
        '503':
          description: Backend unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/models:
    get:
      summary: OpenAI-compatible model list
      operationId: listModels
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                $ref: '../json/models-response.schema.json'
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat completions with llama.cpp extensions
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../json/chat-completion-request.schema.json'
      responses:
        '200':
          description: Chat completion or streaming chunk stream
          content:
            application/json:
              schema:
                $ref: '../json/chat-completion-response.schema.json'
            text/event-stream:
              schema:
                type: string
                description: Server-sent events where each data line contains a chat.completion.chunk JSON object.
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/completions:
    post:
      summary: OpenAI-compatible legacy text completions
      operationId: createCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../json/completion-request.schema.json'
      responses:
        '200':
          description: Text completion
          content:
            application/json:
              schema:
                $ref: '../json/completion-response.schema.json'
  /v1/responses:
    post:
      summary: OpenAI-compatible Responses API subset
      operationId: createResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../json/responses-request.schema.json'
      responses:
        '200':
          description: Response object
          content:
            application/json:
              schema:
                $ref: '../json/responses-response.schema.json'
  /v1/embeddings:
    post:
      summary: OpenAI-compatible embeddings endpoint when llama-server starts with --embeddings
      operationId: createEmbedding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../json/embeddings-request.schema.json'
      responses:
        '200':
          description: Embeddings response
          content:
            application/json:
              schema:
                $ref: '../json/embeddings-response.schema.json'
        '501':
          description: Not enabled unless server starts with --embeddings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Model:
      type: object
      additionalProperties: true
      properties:
        id:
          type: string
        object:
          type: string
          example: model
        created:
          type: integer
        owned_by:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [message]
          properties:
            code:
              oneOf:
                - type: integer
                - type: string
            message:
              type: string
            type:
              type: string